requirements.txt
streamlit
Pillow
pytesseract
pdf2image
EasyOCR
opencv-python-headless # EasyOCR often uses OpenCV
pytest 
pathlib
PyMuPDF
google-generativeai>=0.8.0
python-dotenv
st-gsheets-connection
pandas
sentence-transformers
cosdata-client